classes: Add balena-engine-storage
This class set the BALENA_STORAGE variable at the config parsed event so
that it can be used in an inherit command for native recipes that need
to inherit conditionally on the target's BALENA_STORAGE variable.

This commit can be reversed once `aufs` is deprecated as a supported
storage engine.

Change-type: patch
Signed-off-by: Alex Gonzalez <alexg@balena.io>