balena-net-config: Restart service when config.json changes
This will trigger an update to /run/dnsmasq.servers when config.json
changes for any reason, but importantly when DNS servers are added.

This change also means that balena-net-config no longer remains after
exit and some services that Want or Require it could trigger multiple
executions on each boot or config.json change. This should be safe.

Note that this change alone will not trigger a restart of dnsmasq.

Change-type: patch
Signed-off-by: Kyle Harding <kyle@balena.io>